--- a/tools/cmake/FindOptions.cmake
+++ b/tools/cmake/FindOptions.cmake
@@ -38,7 +38,7 @@ endif()
 
 if(FORCE_AVX512) # some compilers like clang do not automagically define __AVX512__ even when the hardware supports it
 if(NOT MSVC)
-   set (OPT_FLAGS "${OPT_FLAGS} -mbmi2 -mavx512f -mavx512bw -mavx512dq")
+   set (OPT_FLAGS "${OPT_FLAGS} -mbmi2 -mavx512f -mavx512bw -mavx512dq -mavx512vbmi2 -mavx512bitalg -mavx512vpopcntdq")
 else()
    set (OPT_FLAGS "${OPT_FLAGS} /arch:AVX512")
 endif()