Nowadays icelake servers are widely used.  this patch add avx512 optimization for better performance than avx2.  croaring_avx512() will be true if platform support avx512 during runtime dispatch.

Signed-off-by: lifan lin(lifan.lin@intel.com) 
                        hui han(hui.han@intel.com)

here is the benchmark result(platform icelake 6330), 
Not all operations have improvement, but some operations have  obvious performance boost.
avx2                                          
taskset -c 0 benchmarks/add_benchmark         
[cycles/element]
intvlen=1 density=0.200000 order=SHUFFLE
  roaring_bitmap_add():            62.9
  roaring_bitmap_add_many():       85.5
  roaring_bitmap_add_bulk():       65.4
  roaring_bitmap_add_range():      79.4
  roaring_bitmap_remove():        993.9
  roaring_bitmap_remove_range():   81.6
intvlen=4 density=0.200000 order=SHUFFLE
  roaring_bitmap_add():            24.7
  roaring_bitmap_add_many():       23.7
  roaring_bitmap_add_bulk():       23.1
  roaring_bitmap_add_range():      16.6
  roaring_bitmap_remove():        548.9
  roaring_bitmap_remove_range():   26.8
intvlen=16 density=0.200000 order=SHUFFLE
  roaring_bitmap_add():            15.0
  roaring_bitmap_add_many():       10.8
  roaring_bitmap_add_bulk():       12.6
  roaring_bitmap_add_range():       8.1
  roaring_bitmap_remove():        191.8
  roaring_bitmap_remove_range():   15.4
intvlen=64 density=0.200000 order=SHUFFLE
  roaring_bitmap_add():            13.8
  roaring_bitmap_add_many():        8.7
  roaring_bitmap_add_bulk():       11.4
  roaring_bitmap_add_range():       2.0
  roaring_bitmap_remove():         70.2
  roaring_bitmap_remove_range():    3.4

intvlen=1 density=0.200000 order=ASC
  roaring_bitmap_add():             7.5
  roaring_bitmap_add_many():       13.9
  roaring_bitmap_add_bulk():        6.4
  roaring_bitmap_add_range():      27.8
  roaring_bitmap_remove():        606.3
  roaring_bitmap_remove_range():   41.5
intvlen=4 density=0.200000 order=ASC
  roaring_bitmap_add():             7.1
  roaring_bitmap_add_many():        5.3
  roaring_bitmap_add_bulk():        5.8
  roaring_bitmap_add_range():       6.9
  roaring_bitmap_remove():        389.0
  roaring_bitmap_remove_range():   15.1
intvlen=16 density=0.200000 order=ASC
  roaring_bitmap_add():             6.9
  roaring_bitmap_add_many():        3.6
  roaring_bitmap_add_bulk():        5.7
  roaring_bitmap_add_range():       2.7
  roaring_bitmap_remove():        164.3
  roaring_bitmap_remove_range():    8.2
intvlen=64 density=0.200000 order=ASC
  roaring_bitmap_add():             6.9
  roaring_bitmap_add_many():        3.3
  roaring_bitmap_add_bulk():        5.8
  roaring_bitmap_add_range():       0.8
  roaring_bitmap_remove():         66.1
  roaring_bitmap_remove_range():    1.7

intvlen=1 density=0.200000 order=DESC
  roaring_bitmap_add():            28.5
  roaring_bitmap_add_many():       35.1
  roaring_bitmap_add_bulk():       24.5
  roaring_bitmap_add_range():      34.3
  roaring_bitmap_remove():        884.0
  roaring_bitmap_remove_range():   48.6
intvlen=4 density=0.200000 order=DESC
  roaring_bitmap_add():            17.1
  roaring_bitmap_add_many():       14.1
  roaring_bitmap_add_bulk():       14.0
  roaring_bitmap_add_range():       8.6
  roaring_bitmap_remove():        610.2
  roaring_bitmap_remove_range():   18.1
intvlen=16 density=0.200000 order=DESC
  roaring_bitmap_add():            15.3
  roaring_bitmap_add_many():       10.1
  roaring_bitmap_add_bulk():       12.2
  roaring_bitmap_add_range():       4.3
  roaring_bitmap_remove():        182.8
  roaring_bitmap_remove_range():   10.4
intvlen=64 density=0.200000 order=DESC
  roaring_bitmap_add():            16.1
  roaring_bitmap_add_many():       11.2
  roaring_bitmap_add_bulk():       13.8
  roaring_bitmap_add_range():       0.9
  roaring_bitmap_remove():         67.2
  roaring_bitmap_remove_range():    2.0

avx512
taskset -c 0 benchmarks/add_benchmark
[cycles/element]
intvlen=1 density=0.200000 order=SHUFFLE
  roaring_bitmap_add():            64.7
  roaring_bitmap_add_many():       86.4
  roaring_bitmap_add_bulk():       66.5
  roaring_bitmap_add_range():      82.5
  roaring_bitmap_remove():        852.2
  roaring_bitmap_remove_range():   78.1
intvlen=4 density=0.200000 order=SHUFFLE
  roaring_bitmap_add():            25.7
  roaring_bitmap_add_many():       24.6
  roaring_bitmap_add_bulk():       23.0
  roaring_bitmap_add_range():      17.4
  roaring_bitmap_remove():        391.7
  roaring_bitmap_remove_range():   24.5
intvlen=16 density=0.200000 order=SHUFFLE
  roaring_bitmap_add():            15.3
  roaring_bitmap_add_many():       11.7
  roaring_bitmap_add_bulk():       13.1
  roaring_bitmap_add_range():       8.1
  roaring_bitmap_remove():        126.0
  roaring_bitmap_remove_range():   13.1
intvlen=64 density=0.200000 order=SHUFFLE
  roaring_bitmap_add():            14.5
  roaring_bitmap_add_many():        9.5
  roaring_bitmap_add_bulk():       11.6
  roaring_bitmap_add_range():       2.0
  roaring_bitmap_remove():         57.1
  roaring_bitmap_remove_range():    2.9

intvlen=1 density=0.200000 order=ASC
  roaring_bitmap_add():             8.4
  roaring_bitmap_add_many():       15.0
  roaring_bitmap_add_bulk():        6.9
  roaring_bitmap_add_range():      31.6
  roaring_bitmap_remove():        374.2
  roaring_bitmap_remove_range():   36.9
intvlen=4 density=0.200000 order=ASC
  roaring_bitmap_add():             7.7
  roaring_bitmap_add_many():        6.1
  roaring_bitmap_add_bulk():        6.2
  roaring_bitmap_add_range():       7.9
  roaring_bitmap_remove():        226.5
  roaring_bitmap_remove_range():   12.5
intvlen=16 density=0.200000 order=ASC
  roaring_bitmap_add():             7.4
  roaring_bitmap_add_many():        4.4
  roaring_bitmap_add_bulk():        6.0
  roaring_bitmap_add_range():       2.9
  roaring_bitmap_remove():        103.1
  roaring_bitmap_remove_range():    6.2
intvlen=64 density=0.200000 order=ASC
  roaring_bitmap_add():             7.6
  roaring_bitmap_add_many():        4.0
  roaring_bitmap_add_bulk():        6.3
  roaring_bitmap_add_range():       0.9
  roaring_bitmap_remove():         51.6
  roaring_bitmap_remove_range():    1.4

intvlen=1 density=0.200000 order=DESC
  roaring_bitmap_add():            28.8
  roaring_bitmap_add_many():       35.7
  roaring_bitmap_add_bulk():       24.7
  roaring_bitmap_add_range():      38.2
  roaring_bitmap_remove():        752.7
  roaring_bitmap_remove_range():   46.5
intvlen=4 density=0.200000 order=DESC
  roaring_bitmap_add():            17.5
  roaring_bitmap_add_many():       15.0
  roaring_bitmap_add_bulk():       14.4
  roaring_bitmap_add_range():       9.4
  roaring_bitmap_remove():        466.9
  roaring_bitmap_remove_range():   16.6
intvlen=16 density=0.200000 order=DESC
  roaring_bitmap_add():            15.2
  roaring_bitmap_add_many():       10.9
  roaring_bitmap_add_bulk():       12.2
  roaring_bitmap_add_range():       4.4
  roaring_bitmap_remove():        124.5
  roaring_bitmap_remove_range():    8.9
intvlen=64 density=0.200000 order=DESC
  roaring_bitmap_add():            17.1
  roaring_bitmap_add_many():       12.0
  roaring_bitmap_add_bulk():       14.3
  roaring_bitmap_add_range():       0.9
  roaring_bitmap_remove():         57.1
  roaring_bitmap_remove_range():    1.8

avx2
taskset -c 0 benchmarks/bitset_container_benchmark
bitset container benchmarks
set_test(B):  1.50 cycles per operation
get_test(B):  0.96 cycles per operation
bitset_container_cardinality(B):  201.00 cycles per operation
bitset_container_compute_cardinality(B):  0.40 cycles per operation
unset_test(B):  1.51 cycles per operation

 number of values in container = 4096
bitset_container_to_uint32_array(out, Bt, 1234):  0.89 cycles per operation
bitset_container_get bitset_cache_prefetch:  205.69 cycles per operation
bitset_container_get bitset_cache_flush:  327.49 cycles per operation

 number of values in container = 8192
bitset_container_to_uint32_array(out, Bt, 1234):  1.16 cycles per operation
bitset_container_get bitset_cache_prefetch:  208.41 cycles per operation
bitset_container_get bitset_cache_flush:  312.30 cycles per operation

 number of values in container = 16384
bitset_container_to_uint32_array(out, Bt, 1234):  0.60 cycles per operation
bitset_container_get bitset_cache_prefetch:  210.64 cycles per operation
bitset_container_get bitset_cache_flush:  308.05 cycles per operation

 number of values in container = 32768
bitset_container_to_uint32_array(out, Bt, 1234):  0.30 cycles per operation
bitset_container_get bitset_cache_prefetch:  206.42 cycles per operation
bitset_container_get bitset_cache_flush:  309.61 cycles per operation

 number of values in container = 65536
bitset_container_to_uint32_array(out, Bt, 1234):  0.17 cycles per operation
bitset_container_get bitset_cache_prefetch:  212.69 cycles per operation
bitset_container_get bitset_cache_flush:  317.35 cycles per operation


Logical operations (time units per single operation):
bitset_container_and_nocard(B1, B2, BO):  507.00 cycles per operation
bitset_container_and(B1, B2, BO):  534.00 cycles per operation
bitset_container_and_justcard(B1, B2):  353.00 cycles per operation
bitset_container_compute_cardinality(BO):  417.00 cycles per operation
bitset_container_or_nocard(B1, B2, BO):  628.00 cycles per operation
bitset_container_or(B1, B2, BO):  528.00 cycles per operation
bitset_container_or_justcard(B1, B2):  351.00 cycles per operation
bitset_container_compute_cardinality(BO):  419.00 cycles per operation

get_cardinality_through_conversion_to_array(B1):  4.28 cycles per operation

avx512
taskset -c 0 benchmarks/bitset_container_benchmark
bitset container benchmarks
set_test(B):  2.40 cycles per operation
get_test(B):  1.14 cycles per operation
bitset_container_cardinality(B):  203.00 cycles per operation
bitset_container_compute_cardinality(B):  0.31 cycles per operation
unset_test(B):  2.41 cycles per operation

 number of values in container = 4096
bitset_container_to_uint32_array(out, Bt, 1234):  0.96 cycles per operation
bitset_container_get bitset_cache_prefetch:  209.76 cycles per operation
bitset_container_get bitset_cache_flush:  207.77 cycles per operation

 number of values in container = 8192
bitset_container_to_uint32_array(out, Bt, 1234):  0.73 cycles per operation
bitset_container_get bitset_cache_prefetch:  209.72 cycles per operation
bitset_container_get bitset_cache_flush:  208.29 cycles per operation

 number of values in container = 16384
bitset_container_to_uint32_array(out, Bt, 1234):  0.36 cycles per operation
bitset_container_get bitset_cache_prefetch:  210.37 cycles per operation
bitset_container_get bitset_cache_flush:  207.80 cycles per operation

 number of values in container = 32768
bitset_container_to_uint32_array(out, Bt, 1234):  0.21 cycles per operation
bitset_container_get bitset_cache_prefetch:  208.17 cycles per operation
bitset_container_get bitset_cache_flush:  208.69 cycles per operation

 number of values in container = 65536
bitset_container_to_uint32_array(out, Bt, 1234):  0.13 cycles per operation
bitset_container_get bitset_cache_prefetch:  207.84 cycles per operation
bitset_container_get bitset_cache_flush:  208.03 cycles per operation


Logical operations (time units per single operation):
bitset_container_and_nocard(B1, B2, BO):  392.00 cycles per operation
bitset_container_and(B1, B2, BO):  424.00 cycles per operation
bitset_container_and_justcard(B1, B2):  333.00 cycles per operation
bitset_container_compute_cardinality(BO):  322.00 cycles per operation
bitset_container_or_nocard(B1, B2, BO):  392.00 cycles per operation
bitset_container_or(B1, B2, BO):  424.00 cycles per operation
bitset_container_or_justcard(B1, B2):  337.00 cycles per operation
bitset_container_compute_cardinality(BO):  316.00 cycles per operation

get_cardinality_through_conversion_to_array(B1):  4.30 cycles per operation