@lemire 2023-03-13
I agree that this is very exciting work.

Let us make this happen.

@wanweiqiangintel might be interested as well.