@lemire 2023-03-13
@huihan365 When compiling under Visual Studio to a 32-bit target, some code seems to expect croaring_avx512 to be present.