@lemire 2023-03-13
Under the macos build, we are getting illegal instructions which might be related to bitset code I commented: https://github.com/RoaringBitmap/CRoaring/actions/runs/4403958055/jobs/7722282318