@CharlesChen888 2023-03-14
It is exciting to see this PR. Looks like you have implemented AVX512 calculation between bitset containers. Have you done or considered applying AVX512 to array containers?