@lemire 2023-03-14
@CharlesChen888 Note that we can use multiple rounds of PRs to get all the performance out of AVX-512. :-)