@lemire 2023-03-14
@huihan365 

I agree that we do not need to get everything right or perfect at once. You have done a lot of work and I don't want to hold on to the PR to long.

However, I'd like to discuss this point...

> I agree. how about implement in the next release? so this release can apply servers not only icelake but also casalake...

Probably you mean cascade lake... that is, essentially skylake.

Please read this part of the CRoaring documentation...

https://github.com/RoaringBitmap/CRoaring/blob/84fe3c8b0b58111d9daff8256b8399b5d5957968/README.md?plain=1#L592


Notice "so it is not subject to turbo frequency throttling on many-core Intel processors."

As you are no doubt aware, prior to Rocket Lake and Ice Lake, using AVX-512 on Intel processors could lead to severe frequency throttling. I have tried to document it the best I could, based on my experience... https://lemire.me/blog/2018/09/07/avx-512-when-and-how-to-use-these-new-instructions/

Even Intel documentation, at the time, urged engineers to benchmark and assess the benefits. It is fairly complicated...

> On server-class processor, the downclocking is determined on a per-core basis based on the license and the total number of active cores, on the same CPU socket, irrespective of the license of the other cores. That is, to determine the frequency of core under downclocking, you need only to know its license (determined by the type of instructions it runs) and count the number of cores where code is running. Thus you cannot downclock other cores on the same socket, other than the sibling logical core when hyperthreading is used, merely by running heavy and sustained AVX-512 instructions on one core. If you can isolate your heavy numerical work on a few cores (or just one), then the downclocking is limited to these cores. On linux, you can control which cores are running your processes using tools such as taskset or numactl.

Thankfully, [recent Intel servers do not have this issue](https://travisdowns.github.io/blog/2020/08/19/icl-avx512-freq.html).

You can check my answer to @wanweiqiangintel  on a related PR in another project...

https://github.com/simdjson/simdjson/pull/1960

Even Intel would prefer that users do not get bad surprises.

So here is what I propose, if you want to support pre-Ice Lake processor: AVX-512 should be disabled by default on these processors. Maybe a compilation flag (macro) could change this default behaviour.

Only on recent Intel processors (where VBMI2 is available) should AVX-512 be enabled by default.

Now, this makes the engineering more complicated because instead of having to support fallback, AVX2 and AVX-512, are going to have to support fallback, AVX2, AVX-512 and AVX-512-VBMI2. 

I submit to you that it is probably not worth it. It is going to be more effort for relatively little gain in the sense that Intel will quickly move the AVX-512 support to something like Ice Lake or better (that is, VBMI2 + no throttling) across all its server chips.

I am aware that Intel still sells pre-Ice Lake server chips... but I submit to you that it is selling these chips primarily to customers who have little interest for AVX-512.

So my strong suggestion is that we only enable the AVX-512 code when VBMI2 is available. That's where AVX-512 really shines. We won't have to warn users about complicated licensing issues (frequency throttling) and we can do all sorts of nifty tricks.

My recommandation is not that you update your code (e.g., use VPOPCNTDQ and VBMI2) in this PR. It is perfectly acceptable to proceed in steps. However, please consider my point about requiring that the processor has VBMI2. Think about how much simpler it will make future work, and about how it does not require changing out documentation to warn people about frequency throttling.