@lemire 2023-03-14
@huihan365 In effect, short of the small functional changes, I am only suggesting that you check for VBMI2 to enable the AVX-512 kernel.

So I am only suggesting you change a *few* lines.  Your PR is basically good as-is.