@lemire 2023-03-20
@huihan365  Please see https://github.com/RoaringBitmap/CRoaring/pull/451 where I have applied various small fixes, mostly to help portability.