@lemire 2023-03-21
Closing in favour of https://github.com/RoaringBitmap/CRoaring/pull/451